Extract BedrockConfig
