+++ lib/chat_models/chat_anthropic.ex
@@ -56,27 +56,7 @@ defmodule LangChain.ChatModels.ChatAnthropic do
   alias LangChain.Utils
   alias LangChain.Callbacks
   alias LangChain.Utils.BedrockStreamDecoder
-
-  defmodule BedrockConfig do
-    @moduledoc """
-    Configuration for AWS Bedrock.
-    """
-    use Ecto.Schema
-    import Ecto.Changeset
-
-    @primary_key false
-    embedded_schema do
-      # A function that returns a tuple of access_key_id & secret_access_key.
-      field :credentials, :any, virtual: true
-      field :region, :string
-    end
-
-    def changeset(bedrock, attrs) do
-      bedrock
-      |> cast(attrs, [:credentials, :region])
-      |> validate_required([:credentials, :region])
-    end
-  end
+  alias LangChain.Utils.BedrockConfig
 
   @behaviour ChatModel
